requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
#rust based linter
ruff==0.0.284
boto3==1.24.87

# other dependencies
databricks-sql-connector
databricks
pyspark
requests
pandas
numpy
python-dotenv


